// contracts/generated/typescript/models/Collection.ts
export interface Collection {
    /**
     *
     * @type {string}
     * @memberof Collection
     */
    id: string;
    /**
     *
     * @type {string}
     * @memberof Collection
     */
    wikiId: string;
    /**
     *
     * @type {CollectionCollectionTypeEnum}
     * @memberof Collection
     */
    collectionType: CollectionCollectionTypeEnum;
    /**
     *
     * @type {string}
     * @memberof Collection
     */
    title: string;
    /**
     *
     * @type {string}
     * @memberof Collection
     */
    descriptionPageId: string;
    /**
     *
     * @type {CollectionQuery}
     * @memberof Collection
     */
    query: CollectionQuery;
    /**
     *
     * @type {Date}
     * @memberof Collection
     */
    createdAt: Date;
    /**
     *
     * @type {Date}
     * @memberof Collection
     */
    updatedAt: Date;
}


/**
 * @export
 */
export const CollectionCollectionTypeEnum = {
    Manual: 'manual',
    Rule: 'rule',
    Dynamic: 'dynamic'
} as const;
export type CollectionCollectionTypeEnum = typeof CollectionCollectionTypeEnum[keyof typeof CollectionCollectionTypeEnum];